+++ src/sentry/mediators/__init__.py
@@ -1,5 +1,5 @@
+from ..sentry_apps.token_exchange.refresher import Refresher  # noqa: F401
 from .mediator import Mediator  # NOQA
 from .param import Param  # NOQA
 from .token_exchange.grant_exchanger import GrantExchanger  # noqa: F401
-from .token_exchange.refresher import Refresher  # noqa: F401
 from .token_exchange.util import AUTHORIZATION, REFRESH, GrantTypes  # noqa: F401
