+++ src/sentry/mediators/token_exchange/__init__.py
@@ -1,4 +1,4 @@
+from ...sentry_apps.token_exchange.refresher import Refresher  # NOQA
 from .grant_exchanger import GrantExchanger  # NOQA
-from .refresher import Refresher  # NOQA
 from .util import AUTHORIZATION, REFRESH, GrantTypes, token_expiration  # NOQA
 from .validator import Validator  # NOQA
